最大値の選択で再描画
